seprate to new files and small fixes
