+++ src/index.ts
@@ -1,2 +1,8 @@
 export * from './memphis';
-export * from './nest'
+export * from './nest';
+
+export { Consumer } from './consumer';
+export { Message } from './message';
+export { MsgHeaders } from './message-header';
+export { Producer } from './producer';
+export { Station } from './station'
